chore: centralize i18n packages
