+++ apps/marketing/package.json
@@ -18,7 +18,6 @@
     "@documenso/tailwind-config": "*",
     "@documenso/trpc": "*",
     "@documenso/ui": "*",
-    "@formatjs/intl-localematcher": "^0.5.4",
     "@hookform/resolvers": "^3.1.0",
     "@openstatus/react": "^0.0.3",
     "contentlayer": "^0.3.4",
@@ -29,7 +28,6 @@
     "lucide-react": "^0.279.0",
     "luxon": "^3.4.0",
     "micro": "^10.0.1",
-    "negotiator": "^0.6.3",
     "next": "14.0.3",
     "next-auth": "4.24.5",
     "next-axiom": "^1.1.1",
