+++ apps/marketing/src/app/[lang]/(marketing)/blog/[post]/page.tsx
@@ -3,10 +3,11 @@ import Link from 'next/link';
 import { notFound } from 'next/navigation';
 
 import { allBlogPosts } from 'contentlayer/generated';
-import type { Locale } from 'i18n-config';
 import { ChevronLeft } from 'lucide-react';
 import type { MDXComponents } from 'mdx/types';
 import { useMDXComponent } from 'next-contentlayer/hooks';
+
+import type { Locale } from '@documenso/lib/internationalization/i18n-config';
 
 import { CallToAction } from '~/components/(marketing)/call-to-action';
