+++ apps/marketing/tsconfig.json
@@ -21,7 +21,8 @@
     "**/*.ts",
     "**/*.tsx",
     ".next/types/**/*.ts",
-    ".contentlayer/generated"
+    ".contentlayer/generated",
+    "../../packages/lib/internationalization/i18n-config.ts"
   ],
   "exclude": ["node_modules"]
 }
