+++ packages/lib/package.json
@@ -25,6 +25,7 @@
     "@documenso/email": "*",
     "@documenso/prisma": "*",
     "@documenso/signing": "*",
+    "@formatjs/intl-localematcher": "^0.5.4",
     "@next-auth/prisma-adapter": "1.0.7",
     "@noble/ciphers": "0.4.0",
     "@noble/hashes": "1.3.2",
@@ -41,6 +42,7 @@
     "luxon": "^3.4.0",
     "micro": "^10.0.1",
     "nanoid": "^4.0.2",
+    "negotiator": "^0.6.3",
     "next": "14.0.3",
     "next-auth": "4.24.5",
     "oslo": "^0.17.0",
