+++ packages/ui/package.json
@@ -63,7 +63,7 @@
     "lucide-react": "^0.279.0",
     "luxon": "^3.4.2",
     "next": "14.0.3",
-    "pdfjs-dist": "^3.11.174",
+    "pdfjs-dist": "3.11.174",
     "react": "18.2.0",
     "react-colorful": "^5.6.1",
     "react-day-picker": "^8.7.1",
